Generate/map schedule, sessions, speakers in functions (#504)
* Fix updating speakers with sessions

* WIP

* Do not redirect to pages when going through dialogs

* WIP

* Add mapping between sessions and speakers, move adding tags to speakers

* Remove unused function

* Improvements after CR comments

* Improvements after CR comments

* Improvements after CR comments

* Remove webworker
