+++ functions/src/index.js
@@ -1,4 +1,3 @@
-import * as functions from 'firebase-functions';
 import admin from 'firebase-admin';
 
 import saveUserData from './save-user-data';
@@ -7,6 +6,7 @@ import scheduleNotifications from './schedule-notifications';
 import optimizeImages from './optimize-images';
 import mailchimpSubscribe from './mailchimp-subscribe';
 import prerender from './prerender';
+import { scheduleWrite, sessionsWrite, speakersWrite } from './generate-sessions-speakers-schedule';
 
 admin.initializeApp();
 
@@ -17,4 +17,7 @@ export {
   optimizeImages,
   mailchimpSubscribe,
   prerender,
+  scheduleWrite,
+  sessionsWrite,
+  speakersWrite,
 }
