+++ src/mixins/sessions-hoc.html
@@ -36,13 +36,5 @@
         this.dispatch(sessionsActions.fetchList());
       }
     }
-
-    getSessionsBySpeaker(speakerId) {
-      if (this.sessionsBySpeaker) {
-        return this.sessionsBySpeaker[speakerId];
-      }
-
-      return null;
-    }
   };
 </script>
